+++ consensus/dummy/dynamic_fee_accumulator.go
@@ -29,7 +29,7 @@ func CalcACP176GasState(
 		// If the parent block was running with ACP-176, we start with the
 		// parent's fee state.
 		var err error
-		gasAccumulator, err = header.ParseDynamicFeeAccumulator(parent.Extra)
+		gasAccumulator, err = header.ParseDynamicFeeAccumulator(parent.GasLimit, parent.Extra)
 		if err != nil {
 			return header.DynamicFeeAccumulator{}, err
 		}
